requirements.txt
windows-curses
curses
subprocess
tkinter
google-generativeai